Move aggregator support classes out of runners namespace, make private
